Major refactor & clean, implemented threading (still not sure on best practice but it seems fairly simple code wise), added system biasing / persistant clock speed state, stuffed up logging however, need to fix that.

+++ PCSX2_Core/Source/Common/Tables/IOPCoreExceptionsTable/IOPCoreExceptionsTable.cpp
@@ -1,10 +1,8 @@
 #include "stdafx.h"
 
-#include "Common/Global/Globals.h"
-
 #include "Common/Tables/IOPCoreExceptionsTable/IOPCoreExceptionsTable.h"
 
 const IOPCoreExceptionsTable::ExceptionProperties_t * IOPCoreExceptionsTable::getExceptionInfo(const IOPCoreException_t & exception)
 {
-	return &ExceptionProperties[static_cast<u8>(exception)];
+	return &ExceptionProperties[static_cast<int>(exception)];
 }
